// src/domain/models/DataProps.tsx
export interface DataProps {
  collection: string
  collectionDap: string
  destCity: string
  destCountry: string
  destState: string
  destiny: string
  incoterm: string
  oriCity: string
  oriCountry: string
  oriState: string
  origin: string
  originCityName: string
  originCityId: string
  destinationCityName: string
  destinationCityId: string
  idOrigin: string
  idDestination: string
  postalCodeDap: string
  postalCode: string
}